+++ back-end-spring-cloud/high-service/src/main/java/org/example/highservice/mapper/volunteer/MajorMapper.java
@@ -16,5 +16,5 @@ import java.util.List;
 public interface MajorMapper extends BaseMapper<Major> {
     List<Major> searchMajors(@Param("keyword") String keyword,
                              @Param("category") String category,
-                             @Param("universityId") Integer universityId);
+                             @Param("universityId") Long universityId);
 }
